python_lead/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
celery>=5.3,<6.0
redis>=5.0,<6.0
psycopg2-binary>=2.9,<3.0
httpx>=0.25,<1.0
hypothesis>=6.92,<7.0
python-dotenv>=1.0,<2.0
